refactor(analytics): rename support/clip_analysis modules for clearer purpose
Rename support.py → player_motion.py (shared tracking, kinematics, team
helpers, and render utilities) and clip_analysis.py → clip_pipeline.py
(orchestrated per-clip computation producing ClipAnalysis). Update imports
and module docstrings; no behavior changes.
